BAU: export AWS_REGION instead of reconfiguring profile
We were updating ~/.aws/config with the new region in the shell scripts.
This is undesirable, as it can have a weird effect on other scripts that
depend on the 'correct' region being set in the profile.

Instead, we should export AWS_REGION=*whatever* witin the shell script,
to scope the selected region change to that specific operation.