BAU: Add vscode CFN settings + extensions
This configuration will get rid of 99.9% of the errors in the
cfn templates. It just makes it a nicer experience.

The recommended extensions are just to make sure people have these
installed, as they improve the devex of CFN. They are not required,
though - once someone rejects them, they won't be notified again.